Replaced markdown table of standings with rendered images
Telegram does not handle sufficiently large markdown tables well.
The standings table, as printed to a telegram chat prior to this commit,
would commonly overflow onto the next line.

This commit works around this problem by drawing the standings
on a canvas, rendering it to a JPEG image, and sending this image
to the chat.

The bot.send_photo API is used, so pictures are shown inline.